requirements.txt
requests
beautifulsoup4
pydantic
sentence-transformers
scikit-learn
openai
python-dotenv
pytest
numpy
matplotlib